+++ CHANGELOG.md
@@ -40,7 +40,7 @@
 
 * Price numbers are now aligned to the right.
 
-## [Unreleased]
+## [0.0.5] - 2021-04-17 - Bug Fixes
 
 #### Features